renamed _message to #message and moved generateProtocolEncryptedRecordsWrite() method
