+++ tests/utils/test-data-generator.ts
@@ -23,7 +23,9 @@ import type {
 } from '../../src/index.js';
 import {
   DwnInterfaceName,
-  DwnMethodName
+  DwnMethodName,
+  Encryption,
+  KeyDerivationScheme
 } from '../../src/index.js';
 import type { PermissionConditions, PermissionScope, PermissionsGrantMessage, PermissionsRequestMessage, PermissionsRevokeMessage } from '../../src/types/permissions-types.js';
 import type { PrivateJwk, PublicJwk } from '../../src/types/jose-types.js';
@@ -421,6 +423,96 @@ export class TestDataGenerator {
       recordsWrite
     };
   };
+
+  /**
+   * Generates a RecordsWrite message for testing.
+   */
+  public static async generateProtocolEncryptedRecordsWrite(
+    plaintextMessage: Uint8Array,
+    author: Persona,
+    target: Persona,
+    protocolDefinition: ProtocolDefinition,
+    protocolPath: string,
+    protocolContextId: string,
+    protocolContextDerivingRootKeyId: string,
+    protocolContextDerivedPublicJwk: PublicJwk,
+    parentId: string
+  ): Promise<{
+    message: RecordsWriteMessage;
+    dataStream: Readable;
+    recordsWrite: RecordsWrite;
+  }> {
+    // Bob encrypts his initial chat thread to Alice with a randomly generated symmetric key
+    const bobMessageStream = DataStream.fromBytes(plaintextMessage);
+    const dataEncryptionInitializationVector = TestDataGenerator.randomBytes(16);
+    const dataEncryptionKey = TestDataGenerator.randomBytes(32);
+    const bobMessageEncryptedStream = await Encryption.aes256CtrEncrypt(
+      dataEncryptionKey, dataEncryptionInitializationVector, bobMessageStream
+    );
+    const bobMessageEncryptedBytes = await DataStream.toBytes(bobMessageEncryptedStream);
+
+    // Bob generates a RecordsWrite for the encrypted message
+    const protocolPathSegments = protocolPath.split('/');
+    const recordType = protocolPathSegments[protocolPathSegments.length - 1];
+    const { message, dataStream, recordsWrite } = await TestDataGenerator.generateRecordsWrite(
+      {
+        author,
+        protocol   : protocolDefinition.protocol,
+        protocolPath,
+        contextId  : protocolContextId, // TODO:
+        parentId, // TODO:
+        schema     : protocolDefinition.types[recordType].schema,
+        dataFormat : protocolDefinition.types[recordType].dataFormats![0],
+        data       : bobMessageEncryptedBytes
+      }
+    );
+
+    // Bob prepares the symmetric key encryption input for protocol-path derived public key
+    let protocolSegment = protocolDefinition.structure;
+    for (const pathSegment of protocolPathSegments) {
+      protocolSegment = protocolSegment[pathSegment];
+    }
+
+    const protocolPathDerivedPublicJwk = protocolSegment.$encryption?.publicKeyJwk;
+    const protocolPathDerivedKeyEncryptionInput = {
+      publicKeyId      : target.keyId, // TODO: should come from the $encryption property
+      publicKey        : protocolPathDerivedPublicJwk!,
+      derivationScheme : KeyDerivationScheme.ProtocolPath
+    };
+
+    // Bob prepares the symmetric key encryption input for protocol-context derived public key
+    // const bobRootPrivateKey: DerivedPrivateJwk = {
+    //   rootKeyId         : author.keyId,
+    //   derivationScheme  : KeyDerivationScheme.ProtocolContext,
+    //   derivedPrivateKey : author.keyPair.privateJwk
+    // };
+
+    // const contextId = await RecordsWrite.getEntryId(author.did, message.descriptor);
+    // const contextDerivationPath = Records.constructKeyDerivationPathUsingProtocolContextScheme(contextId);
+    // const protocolContextDerivedPublicJwk = await HdKey.derivePublicKey(bobRootPrivateKey, contextDerivationPath);
+
+    const protocolContextDerivedKeyEncryptionInput = {
+      publicKeyId      : protocolContextDerivingRootKeyId,
+      publicKey        : protocolContextDerivedPublicJwk,
+      derivationScheme : KeyDerivationScheme.ProtocolContext
+    };
+
+    // final encryption input
+    const encryptionInput: EncryptionInput = {
+      initializationVector : dataEncryptionInitializationVector,
+      key                  : dataEncryptionKey,
+      keyEncryptionInputs  : [protocolPathDerivedKeyEncryptionInput, protocolContextDerivedKeyEncryptionInput]
+    };
+
+    await recordsWrite.encryptSymmetricEncryptionKey(encryptionInput);
+    await recordsWrite.sign(Jws.createSignatureInput(author));
+
+    // // Bob writes the encrypted email to Alice's DWN
+    // const bobWriteReply = await dwn.processMessage(target.did, message, dataStream);
+    // expect(bobWriteReply.status.code).to.equal(202);
+
+    return { message, dataStream: dataStream!, recordsWrite };
+  }
 
   /**
    * Generates a valid RecordsWrite that modifies the given an existing write.
